refactor: Moved Participant Class file to shared model folder
