+++ src/ParticipantManagementService/GetParticipant/ParticipantRepository.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using NHS.ServiceInsights.Model;
 
 namespace GetParticipant
 {
